Role selection screen now server rendered
