+++ apps/website/app/guild/[guildId]/layout.tsx
@@ -1,12 +1,13 @@
 import * as context from 'next/headers';
 import Image from 'next/image';
 import Link from 'next/link';
-import { notFound } from 'next/navigation';
+import { notFound, redirect } from 'next/navigation';
 import { Button } from 'app/_components/ui/button';
 import { Separator } from 'app/_components/ui/separator';
 import { CDNRoutes, ImageFormat, RouteBases } from 'discord-api-types/v10';
 import { Settings, Users } from 'lucide-react';
 import rscApi from 'utils/rsc';
+import { TRPCError } from '@trpc/server';
 
 type Props = {
   children?: React.ReactNode;
@@ -18,7 +19,18 @@ type Props = {
 const GuildLayout = async ({ children, params: { guildId } }: Props) => {
   const api = await rscApi(context);
 
-  const guild = (await api.guild.get({ guildId }))?.discord;
+  const guild = (await api.guild.get({ guildId })
+    .catch(err => {
+      if(err instanceof TRPCError) {
+        if (err.code === 'NOT_FOUND') {
+          notFound();
+        }
+
+        if (err.code === 'UNAUTHORIZED') {
+          redirect('/login/discord');
+        }
+      }
+    }))?.discord;
 
   const icon =
     guild?.icon &&
@@ -49,7 +61,7 @@ const GuildLayout = async ({ children, params: { guildId } }: Props) => {
           </span>
           {guild?.name}
         </h1>
-        <div className="flex-1 rounded-md bg-primary-100 p-2 dark:bg-primary-900">
+        <div className="rounded-md bg-primary-100 p-2 dark:bg-primary-900">
           <div className="py-2">
             <Button asChild className="w-full justify-start" variant="outline">
               <Link href={`/guild/${guildId}/roles`}>
